requirements.txt
anthropic>=0.9.0
openai==1.12.0
google-generativeai==0.3.2
PyPDF2==3.0.1
pandas==2.2.1
tqdm==4.66.2
python-dotenv==1.0.1 